CentOS 6.5 (Latest) Compatibility
Updated the Makefiles to correctly build on a 64-bit Linux system.  Currently, the Makefile will build cheshire on a system running CentOS 6.5.  A copy of the makefile is available in Makefile.centos65.

Other makefiles corrected to build and clean the system correctly.  Code from previous commits WILL NOT compile on a CentOS system.